+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "Project Timer" extension will be documented in this file.
 
+## [Unreleased]
+### Added
+- Support to match project by the first remote url of current repository. 
+
 ## [0.1.1] - 2026-2-20
 ### Added
 - Storage cache system to improve performance.